+++ adapters/DTExchange/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## DT Exchange iOS Mediation Adapter Changelog
 
+#### Next version
+- Added watermark for bidding ads.
+
 #### [Version 8.4.4.0](https://dl.google.com/googleadmobadssdk/mediation/ios/dtexchange/DTExchangeAdapter-8.4.4.0.zip)
 - Verified compatibility with DT Exchange SDK version 8.4.4.
 - Added support for forwarding the `tagForChildDirectedTreatment` and `tagForUnderAgeOfConsent` Google Mobile Ads SDK parameters to the DT Exchange SDK.